[feat] MindSharePostsRequest pageSize int로 변경

+++ src/main/java/com/jydev/mindtravel/service/mind/share/model/MindSharePostsRequest.java
@@ -7,6 +7,6 @@ import lombok.Getter;
 @Getter
 public class MindSharePostsRequest {
     private Long pageOffset;
-    private Long pageSize;
+    private Integer pageSize;
     private MindSharePostCategory category;
 }
